requirements.txt
Django~=4.1.5
django-embed-video~=1.4.8